Add chapter controller/service/repo/routes
Introduce full chapter module: add ChapterController, ChapterService, ChapterRepository and factory, plus chapter routes. Add Zod schemas for creating, deleting and listing chapters (with pagination) and a create.volume schema. Update interfaces (IChapterRepository, IChapterService, IVolumeRepository) to reflect new DTOs and string IDs. Service enforces duplicate order check and throws ConflictError; repository implements pagination and duplicate control. Adjust entity column types: Chapter.order -> int, Volume.order -> float. Add reqFloat helper and register chapter routes in main router.

+++ src/entities/Chapter.ts
@@ -23,7 +23,7 @@ export class Chapter {
   @Column({ type: "text" })
   content!: string;
 
-  @Column({ type: "float" })
+  @Column({ type: "int" })
   order!: number;
 
   @CreateDateColumn()
